Update README.md and move AppConstant under utils.
