+++ app/src/main/java/com/sachin_singh_dighan/newsapp/NewsApplication.kt
@@ -8,6 +8,7 @@ import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.NetworkType
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkManager
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.utils.TimeUtil
 import com.sachin_singh_dighan.newsapp.worker.TopHeadlinesSyncWorker
 import dagger.hilt.android.HiltAndroidApp
