+++ app/src/main/java/com/sachin_singh_dighan/newsapp/data/repository/languageselection/LanguageSelectionRepository.kt
@@ -1,6 +1,6 @@
 package com.sachin_singh_dighan.newsapp.data.repository.languageselection
 
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.data.model.languageselection.LanguageData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
