+++ app/src/main/java/com/sachin_singh_dighan/newsapp/ui/countryselection/CountrySelectionViewModel.kt
@@ -1,7 +1,7 @@
 package com.sachin_singh_dighan.newsapp.ui.countryselection
 
 import androidx.lifecycle.viewModelScope
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.data.model.countryselection.CountrySelection
 import com.sachin_singh_dighan.newsapp.data.repository.countryselection.CountrySelectionRepository
 import com.sachin_singh_dighan.newsapp.ui.base.BaseViewModel
