+++ app/src/main/java/com/sachin_singh_dighan/newsapp/ui/searchnews/SearchNewsViewModel.kt
@@ -1,7 +1,7 @@
 package com.sachin_singh_dighan.newsapp.ui.searchnews
 
 import androidx.lifecycle.viewModelScope
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.data.model.topheadline.ApiArticle
 import com.sachin_singh_dighan.newsapp.data.repository.searchnews.SearchNewsRepository
 import com.sachin_singh_dighan.newsapp.ui.base.BaseViewModel
