+++ app/src/main/java/com/sachin_singh_dighan/newsapp/worker/TopHeadlineSyncScheduler.kt
@@ -7,7 +7,7 @@ import androidx.work.NetworkType
 import androidx.work.OneTimeWorkRequest
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import java.util.Calendar
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
