+++ app/src/test/java/com/sachin_singh_dighan/newsapp/data/repository/mainscreen/MainRepositoryTest.kt
@@ -1,7 +1,7 @@
 package com.sachin_singh_dighan.newsapp.data.repository.mainscreen
 
 import app.cash.turbine.test
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.data.model.mainscreen.MainSection
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.test.runTest
