+++ app/src/test/java/com/sachin_singh_dighan/newsapp/ui/languageselection/LanguageSelectionViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.sachin_singh_dighan.newsapp.ui.languageselection
 
 import app.cash.turbine.test
-import com.sachin_singh_dighan.newsapp.AppConstant
+import com.sachin_singh_dighan.newsapp.utils.AppConstant
 import com.sachin_singh_dighan.newsapp.data.model.languageselection.LanguageData
 import com.sachin_singh_dighan.newsapp.data.repository.languageselection.LanguageSelectionRepository
 import com.sachin_singh_dighan.newsapp.ui.common.UiState
